Return early when the value is not a string
The subsequent logic calls `ltrim` which expects the input to be a string.
If this is not the case, there will be an error thrown about this.

At least one code-path that leads to this situation is via the following line: https://github.com/WordPress/wordpress-develop/blob/master/src/wp-includes/update.php#L197

The response of WordPress.org for core does not return a value for all "package" entries. Thus passing `false` to the `esc_url` array_map.

```php
Array
(
    [full] => https://downloads.wordpress.org/release/wordpress-5.6.1.zip
    [no_content] => https://downloads.wordpress.org/release/wordpress-5.6.1-no-content.zip
    [new_bundled] => https://downloads.wordpress.org/release/wordpress-5.6.1-new-bundled.zip
    [partial] =>
    [rollback] =>
)
```